+++ src/Markdown.MAML/Model/MAML/MamlSyntax.cs
@@ -11,11 +11,13 @@ namespace Markdown.MAML.Model.MAML
         public MamlSyntax()
         {
             // default for parameter set names is __AllParameterSets
-            ParameterSetName = "__AllParameterSets";
+            //commented out for future consideration.
+            //ParameterSetName = "__AllParameterSets";
         }
-
-
+        
         public string ParameterSetName { get; set; }
+
+        public bool IsDefault { get; set; }
 
         public List<MamlParameter> Parameters { get { return _parameters; } }